+++ src/local_pathfinding/local_pathfinding/global_path_storage/main_global_path.csv
@@ -1,11 +1,3 @@
 latitude,longitude
-49.43243408203125,-123.47017669677734
-49.415531158447266,-123.43785095214844
-49.39862060546875,-123.4055404663086
-49.3817024230957,-123.37325286865234
-49.364776611328125,-123.34098815917969
-49.34783935546875,-123.30874633789062
-49.33089065551758,-123.27652740478516
-49.31393814086914,-123.24433135986328
-49.296974182128906,-123.21215057373047
-49.279998779296875,-123.18000030517578
+48.158390045166016,-130.25390625
+48.121368408203125,-137.02294921875
